+++ webtask-templates/common.json
@@ -454,30 +454,6 @@
           "value": "ublkdu"
         }
       ]
-    },
-    "SPLUNK_URL": {
-      "description": "Splunk URL - this is your Splunk HTTP Collector Endpoint",
-      "required": true
-    },
-    "SPLUNK_TOKEN": {
-      "description": "Splunk Token - this is your Splunk Token",
-      "required": true,
-      "type": "password"
-    },
-    "SPLUNK_COLLECTOR_PORT": {
-      "description": "HTTP Collector Port",
-      "required": true,
-      "default": 8088
-    },
-    "SPLUNK_COLLECTOR_PATH": {
-      "description": "HTTP Collector Path (Endpoint)",
-      "required": false,
-      "default": "/services/collector/event/1.0"
     }
-  },
-  "title": "Auth0 Logs to Splunk",
-  "name": "auth0-logs-to-splunk",
-  "version": "0.1.0",
-  "description": "This extension will take all of your Auth0 logs and export them to Splunk",
-  "logoUrl": "https://cdn.auth0.com/extensions/auth0-logs-to-splunk/assets/logo.svg"
+  }
 }
